Clean up old classes, restructure vision code
